pyproject.toml
[project]
name = "psychology-experiment-reproduce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.2",
    "pandas>=2.2.2",
    "langgraph>=1.0.4",
    "loguru>=0.7.3",
    "manim>=0.19.1",
    "openhands-agent-server>=1.4.0",
    "openhands-sdk>=1.4.0",
    "openhands-tools>=1.4.0",
    "openhands-workspace>=1.4.0",
    "openpyxl>=3.1.5",
    "promptbind>=0.2.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
